+++ Rona/Param_Exam.php
@@ -5,7 +5,7 @@
  * @copyright Copyright (c) 2018 Ryan Whitman (https://ryanwhitman.com)
  * @license https://opensource.org/licenses/MIT
  * @link https://github.com/RyanWhitman/ronaphp
- * @version 1.4.0
+ * @version 1.5.0
  */
 
 namespace Rona;
@@ -121,7 +121,7 @@ class Param_Exam {
 		// Create an array to hold the failed data.
 		$failed_data = [];
 
-		// Run thru the exams. Using an infinite loop allows exams to be dynamically added with the callback methods.
+		// Run through the exams. Using an infinite loop allows exams to be dynamically added with the callback methods.
 		while (1) {
 
 			// If no more exams exist, break the infinite loop.